requirements.txt
fastapi[all]==0.110.0
supabase==2.4.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg2-binary
